requirements/base.txt
Django==5.0.6
gunicorn==22.0.0
django-model-utils==4.5.1